Fix build after Lunchbox explicit ctor changes

+++ tests/objectMaxVersions.cpp
@@ -65,7 +65,7 @@ public:
     virtual void run()
         {
             lunchbox::sleep( 110 );
-            _object.sync( 3 );
+            _object.sync( uint128_t(3) );
             TESTINFO( _object.getVersion() == 3, _object.getVersion( ));
         }
